URLencode and decode username strings for issue #6
We weren't correctly encoding the username string in the eavesdrop.fm front-end, or parsing it this way in the webhook worker. If your scrobble events are currently not working, try setting up the webhook again from the website, or editing your existing webhook manually by replacing any space characters with `%20%`.